🚚 rename: InfoPart 위치 이동 및 h2 태그로 변경 #58
