+++ src/components/PayModal.tsx
@@ -9,6 +9,7 @@ import Button from "@/common/components/atoms/Button";
 import Loading from "@/common/components/atoms/Loading";
 import successToast from "@/common/utils/successToast";
 import DetailedTicket from "@/common/components/molecules/DetailedTicket";
+import InfoPart from "@/common/components/molecules/InfoPart";
 
 interface PayModalProps {
   isOpen: boolean;
@@ -93,24 +94,5 @@ export default function PayModal({
         )}
       </section>
     </Modal>
-  );
-}
-
-function InfoPart({
-  heading,
-  content,
-  isRight,
-  isSeat,
-}: {
-  heading: string;
-  content: string;
-  isRight?: true;
-  isSeat?: true;
-}) {
-  return (
-    <div className={`flex flex-col ${isSeat && "w-3/4"} ${isRight && "items-end"}`}>
-      <div className="text-[16px] text-text-secondary">{heading}</div>
-      <div className={`text-[16px] ${isSeat && "break-keep"}`}>{content}</div>
-    </div>
   );
 }
